// apps/frontend2/src/api/models/createEducationRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { CreateEducationRequestField } from './createEducationRequestField';
import type { CreateEducationRequestEndDate } from './createEducationRequestEndDate';
import type { CreateEducationRequestDescription } from './createEducationRequestDescription';

export interface CreateEducationRequest {
  /** @minLength 1 */
  school: string;
  start_date: string;
  degree: string;
  field?: CreateEducationRequestField;
  end_date?: CreateEducationRequestEndDate;
  description?: CreateEducationRequestDescription;
}