// apps/frontend2/src/api/models/profileEducationDTO.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileEducationDTODegree } from './profileEducationDTODegree';
import type { ProfileEducationDTOField } from './profileEducationDTOField';
import type { ProfileEducationDTOEndDate } from './profileEducationDTOEndDate';
import type { ProfileEducationDTODescription } from './profileEducationDTODescription';

export interface ProfileEducationDTO {
  id: number;
  school: string;
  degree: ProfileEducationDTODegree;
  field: ProfileEducationDTOField;
  start_date: string;
  end_date: ProfileEducationDTOEndDate;
  description: ProfileEducationDTODescription;
  created_at: string;
  updated_at: string;
}