// apps/frontend2/src/api/models/updateEducationRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * FastAPI
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateEducationRequestSchool } from './updateEducationRequestSchool';
import type { UpdateEducationRequestDegree } from './updateEducationRequestDegree';
import type { UpdateEducationRequestField } from './updateEducationRequestField';
import type { UpdateEducationRequestStartDate } from './updateEducationRequestStartDate';
import type { UpdateEducationRequestEndDate } from './updateEducationRequestEndDate';
import type { UpdateEducationRequestDescription } from './updateEducationRequestDescription';

export interface UpdateEducationRequest {
  school?: UpdateEducationRequestSchool;
  degree?: UpdateEducationRequestDegree;
  field?: UpdateEducationRequestField;
  start_date?: UpdateEducationRequestStartDate;
  end_date?: UpdateEducationRequestEndDate;
  description?: UpdateEducationRequestDescription;
}